+++ server/impl/src/main/java/com/walmartlabs/concord/server/console/PasswordValidationResponse.java
@@ -1,0 +1,4 @@
+package com.walmartlabs.concord.server.console;
+
+public class PasswordValidationResponse {
+}